Implemented changes discussed in reveiw comments
