+++ android/assets/jsons/translations/template.properties
@@ -921,12 +921,6 @@ Your [ourUnit] captured an enemy [theirUnit]! =
 Your [ourUnit] plundered [amount] [Stat] from [theirUnit] = 
 We have captured a barbarian encampment and recovered [goldAmount] gold! = 
 An enemy [unitType] has joined us! = 
-After an unknown civilization entered the [eraName], we have recruited [spyName] as a spy! = 
-We have recruited [spyName] as a spy! = 
-A Spy from [civilization] stole the Technology [techName] from [cityName]! = 
-An unidentifeid spy stole the Technology [techName] from [cityName]! = 
-Your spy [name] stole the Technology [techName] from [cityName]! = 
-Your spy [name] cannot steal any more techs from [civilization] as we've already researched all the technology they know! = 
 [unitName] can be promoted! = 
 
 # This might be needed for a rewrite of Germany's unique - see #7376
@@ -1655,8 +1649,14 @@ Spy =
 Spy Hideout = 
 Spy present = 
 Move = 
-Stealing Tech = 
-Surveiying City = 
+
+After an unknown civilization entered the [eraName], we have recruited [spyName] as a spy! = 
+We have recruited [spyName] as a spy! = 
+A spy from [civilization] stole the Technology [techName] from [cityName]! = 
+An unidentified spy stole the Technology [techName] from [cityName]! = 
+Your spy [name] stole the Technology [techName] from [cityName]! = 
+Your spy [name] cannot steal any more techs from [civilization] as we've already researched all the technology they know! = 
+
 
 # Promotions
